requirements.txt
streamlit==1.36.0
pandas>=2.0.0
pyyaml>=6.0.0
python-pptx>=0.6.22
reportlab>=4.0.0
